+++ src/event-selector/interfaces.ts
@@ -1,0 +1,4 @@
+import { EventPickerProps } from '../event-picker';
+import { SelectorProps } from '../selector';
+
+export interface EventSelectorProps extends EventPickerProps, Omit<SelectorProps, 'dropdownRender' | 'valueRender'> {}