+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## [21.4.3-alpha.1](https://github.com/growingio/gio-design-pro/compare/v21.4.2...v21.4.3-alpha.1) (2021-04-27)
+
+
+### Bug Fixes
+
+* **filter-picker:** 修复日期类型,返回值为数字,以及添加op配置字段 ([#214](https://github.com/growingio/gio-design-pro/issues/214)) ([dcb5a7d](https://github.com/growingio/gio-design-pro/commit/dcb5a7d5fc707f2b2f6a35ed4da320db207aaab3))
+* **operation-menu:** 修复因为gio-design版本升级导致的样式问题 ([#212](https://github.com/growingio/gio-design-pro/issues/212)) ([da7c3e7](https://github.com/growingio/gio-design-pro/commit/da7c3e72fa44086c776dbd399005f733811b9bdc))
+
+
+### Features
+
+* **event-picker:** add component event-picker ([41dfcc3](https://github.com/growingio/gio-design-pro/commit/41dfcc318396404acc4c51cefe3fca719bcfb99a))
+* **event-picker:** add component event-picker ([1417fdc](https://github.com/growingio/gio-design-pro/commit/1417fdc7113ae067d980c15dc5dd4fe6a47d400a))
+
+
+
 ## [21.4.2](https://github.com/growingio/gio-design-pro/compare/v21.4.1...v21.4.2) (2021-04-22)